GOV-11320 - added versions array back to GetMetricByuid in metrics controller, GOV-11362 - fixed single to double conversion error for Threshold field

+++ d360.core/entities/Metric/MetricAssetVersionCondition.cs
@@ -20,7 +20,7 @@ namespace d360.core.entities.Metric
         public int Position { get; set; }
 
         [DataMember]
-        public float? Threshold { get; set; }
+        public double? Threshold { get; set; }
 
         [DataMember]
         public decimal? Weight { get; set; }
